Complete project rebase to a Vaadin 10 starter base.
I borked something pretty badly along the way. I think I was mixing vaadin 8 and vaadin 11, which are incompatible as far as i can tell.

Figured it best to throw the baby out with the bathwater and pull a starter project from vaadin and start the project fresh from there.

Vaadin designer should now be useable, as well as the features of vaadin 10.