many changes. much progress.
folded the dice class into the player class. Dice class was uneccessary.

changes to the function of the player. for the time being, the players are hard coded.

changes to start and goal pools. each is now a set. player now instantiates  both upon its own instatiation.

board is now defined as a singleton.